refactor(cli): split command groups into domain modules
Move CLI implementations into docs/project/self/main modules, point the
script entrypoint at main_app, and add confirmation friction for full PAT
display. Update tests and document the breaking import contract in ADRs.

Refs #43
