--- src/siesta/cli.py
@@ -1,1377 +1,0 @@
-# Copyright 2025 Entalpic
-"""
-Source code for the ``siesta`` Command-Line Interface (CLI).
-
-Learn how to use with:
-
-.. code-block:: bash
-
-    # Top-level commands, add --help to any of them
-    # to understand what they do and their options their options.
-
-    $ siesta # shows help
-
-    $ siesta docs # shows help
-    $ siesta docs init
-    $ siesta docs open
-    $ siesta docs build
-    $ siesta docs update
-    $ siesta docs watch
-
-    $ siesta project # shows help
-    $ siesta project quickstart
-    $ siesta project setup-tests
-    $ siesta project tree
-
-    $ siesta self # shows help
-    $ siesta self version
-    $ siesta self update  # or: siesta self upgrade
-    $ siesta self set-github-pat
-    $ siesta self show-github-pat
-    $ siesta self show-deps
-
-You can also refer to the :ref:`siesta-cli-tutorial` for more information.
-
-.. note::
-
-    siesta checks for updates in the background (cached for 24 hours). To change
-    the frequency, set the ``SIESTA_UPDATE_CHECK_HOURS`` environment variable.
-    Set it to ``"false"`` or ``"-1"`` to disable automatic checks.
-
-"""
-
-import builtins
-import getpass
-import os
-import platform
-import subprocess
-import sys
-import time
-from importlib import metadata
-from pathlib import Path
-from shutil import get_terminal_size, rmtree
-from textwrap import dedent
-from typing import Annotated, Optional
-
-from cyclopts import App, Parameter
-from gitignore_parser import parse_gitignore
-from rich import print as rprint
-from watchdog.observers import Observer
-
-from siesta.completions import Shell, detect_current_shell, is_completion_installed
-from siesta.utils.common import (
-    get_project_name,
-    load_deps,
-    logger,
-    resolve_path,
-    run_command,
-    write_or_update_pre_commit_file,
-)
-from siesta.utils.config import CLI_DEFAULTS
-from siesta.utils.docs import (
-    AutoBuildDocs,
-    copy_boilerplate,
-    has_python_files,
-    install_dependencies,
-    make_empty_folders,
-    overwrite_docs_files,
-    update_conf_py,
-    write_rtd_config,
-)
-from siesta.utils.github import (
-    get_latest_commit_info,
-    get_latest_github_release_version,
-    get_user_pat,
-)
-from siesta.utils.project import (
-    add_ipdb_as_debugger,
-    write_gitignore,
-    write_test_actions_config,
-    write_tests_infra,
-)
-from siesta.utils.self import (
-    compare_versions,
-    get_installation_method,
-    get_latest_version,
-    get_update_command,
-    get_update_message,
-    start_background_update_check,
-    update_siesta,
-)
-from siesta.utils.tree import make_labeled_tree
-
-# Main app
-app = App(
-    help=dedent(
-        f"""
-    Siesta Is Entalpic'S Terminal Assistant ({metadata.version("siesta")})
-    
-    A set of CLI tools to help you with good practices in Python development at Entalpic.
-
-    Upgrade with ``$ siesta self update``.
-
-    See Usage instructions in the online docs: https://entalpic-siesta.readthedocs-hosted.com/en/latest/autoapi/siesta/.
-    """.strip()
-    ),
-)
-""":py:class:`cyclopts.App`: The main CLI application."""
-
-docs_app = App(
-    name="docs",
-    help=dedent(
-        """
-        Initialize, build and watch a Sphinx documentation project with standard Entalpic config.
-
-        Upgrade with ``$ siesta self update``.
-
-        See Usage instructions in the online docs: https://entalpic-siesta.readthedocs-hosted.com/en/latest/autoapi/siesta.
-
-        """.strip(),
-    ),
-)
-""":py:class:`cyclopts.App`: The app for the ``siesta docs`` sub-command."""
-
-project_app = App(
-    name="project",
-    help=dedent(
-        """
-        Initialize and manage Python projects with standard Entalpic config.
-
-        Upgrade with ``$ siesta self update``.
-
-        See Usage instructions in the online docs: https://entalpic-siesta.readthedocs-hosted.com/en/latest/autoapi/siesta.
-
-        """.strip(),
-    ),
-)
-""":py:class:`cyclopts.App`: The app for the ``siesta project`` sub-command."""
-
-self_app = App(
-    name="self",
-    help=dedent(
-        """
-        Manage siesta itself and configure settings.
-
-        Includes version checking, updates, GitHub PAT configuration, and dependency info.
-
-        """.strip(),
-    ),
-)
-""":py:class:`cyclopts.App`: The app for the ``siesta self`` sub-command."""
-
-app.command(docs_app)
-app.command(project_app)
-app.command(self_app)
-
-# ============================================================================
-# Tab completions sub-app (under `siesta self tab-completions`)
-# ============================================================================
-
-tab_completions_app = App(
-    name="tab-completions",
-    help="Command group — manage shell tab completions (requires a subcommand).",
-)
-self_app.command(tab_completions_app)
-
-
-def _resolve_shell(shell: Shell | None) -> Shell:
-    """Return ``shell`` if given, otherwise detect and return the current shell.
-
-    Parameters
-    ----------
-    shell : Shell | None
-        Explicitly provided shell value, or ``None`` to auto-detect.
-
-    Returns
-    -------
-    Shell
-        The resolved shell (``"bash"`` or ``"zsh"``).
-    """
-    if shell is not None:
-        return shell
-    detected = detect_current_shell()
-    if detected is None:
-        rprint(
-            "[red]Cannot detect current shell. "
-            "Please specify --shell bash or --shell zsh.[/red]"
-        )
-        sys.exit(1)
-    return detected
-
-
-@tab_completions_app.command
-def install(
-    *,
-    shell: Shell | None = None,
-    add_to_startup: bool = True,
-) -> None:
-    """Install tab completions for the current (or specified) shell.
-
-    Writes managed hook and static completion script files to the
-    shell-specific XDG config path, and optionally adds a source line to the
-    shell RC file.
-
-    Parameters
-    ----------
-    shell : {'bash', 'zsh'} | None, optional
-        Target shell. Defaults to the current shell detected from ``$SHELL``.
-        Required when the current shell cannot be detected.
-    add_to_startup : bool, optional
-        When ``True`` (default), append a source line to ``~/.bashrc`` or
-        ``~/.zshrc`` so completions activate in every new shell session.
-        Use ``--no-add-to-startup`` to write the files without editing the RC
-        file.
-    """
-    from siesta.completions import install_managed_completion
-
-    resolved = _resolve_shell(shell)
-    path = install_managed_completion(
-        shell=resolved,
-        generate_fn=app.generate_completion,
-        add_to_startup=add_to_startup,
-    )
-    rprint(f"[green]Managed completion installed to {path}[/green]")
-    if add_to_startup:
-        rc = ".bashrc" if resolved == "bash" else ".zshrc"
-        rprint(
-            f"[dim]Restart your shell or run `source ~/{rc}` "
-            "to activate completions now.[/dim]"
-        )
-
-
-@tab_completions_app.command
-def show(*, shell: Shell | None = None) -> None:
-    """Print the tab completion script to stdout (without installing).
-
-    Outputs the raw completion script for the given shell; suitable for
-    manual sourcing or inspection.
-
-    Parameters
-    ----------
-    shell : {'bash', 'zsh'} | None, optional
-        Target shell. Defaults to the current shell detected from ``$SHELL``.
-        Required when the current shell cannot be detected.
-    """
-    resolved = _resolve_shell(shell)
-    builtins.print(app.generate_completion(shell=resolved), end="")
-
-
-@tab_completions_app.command
-def where(*, shell: Shell | None = None, simple: bool = False) -> None:
-    """Display where managed completion files are stored.
-
-    By default shows a table with the path, existence status, and purpose of
-    each managed file. Use ``--simple`` for plain-text output (one path per
-    line), which is handy for scripting.
-
-    Parameters
-    ----------
-    shell : {'bash', 'zsh'} | None, optional
-        Target shell. Defaults to the current shell detected from ``$SHELL``.
-        Required when the current shell cannot be detected.
-    simple : bool, optional
-        When ``True``, print one ``<label>: <path>`` line per file with no
-        table formatting or descriptions.
-    """
-    from siesta.completions import (
-        managed_completion_paths,
-        resolve_cli_executable,
-        stable_exec_id,
-    )
-
-    resolved = _resolve_shell(shell)
-    exec_path = resolve_cli_executable()
-    exec_id = stable_exec_id(exec_path)
-    paths = managed_completion_paths(shell=resolved, exec_id=exec_id)
-
-    _ENTRIES = [
-        (
-            "base_dir",
-            "Base directory",
-            "Root folder for all managed completion files for this shell.",
-        ),
-        (
-            "hook_file",
-            "Hook file",
-            f"Sourced by ~/{'.bashrc' if resolved == 'bash' else '.zshrc'} "
-            "at shell startup. Sources the static script and wraps the CLI command.",
-        ),
-        (
-            "static_file",
-            "Static completion script",
-            "Cached completion script generated by the CLI framework. "
-            "Sourced instantly at startup — no Python spawned. Its mtime "
-            "tracks the installed CLI version for staleness detection.",
-        ),
-    ]
-
-    if simple:
-        for key, label, _ in _ENTRIES:
-            p: Path = paths[key]
-            exists = "[green]exists[/green]" if p.exists() else "[dim]missing[/dim]"
-            rprint(f"[bold]{label}:[/bold] {p}  ({exists})")
-        return
-
-    from rich.table import Table
-
-    table = Table(
-        title=f"Managed completion paths — [cyan]{resolved}[/cyan]",
-        show_lines=True,
-    )
-    table.add_column("File", style="bold", no_wrap=True)
-    table.add_column("Status", no_wrap=True)
-    table.add_column("Path", no_wrap=True)
-    table.add_column("Description")
-
-    for key, label, description in _ENTRIES:
-        p = paths[key]
-        status = "[green]exists[/green]" if p.exists() else "[dim]missing[/dim]"
-        table.add_row(label, status, str(p), description)
-
-    rprint(table)
-
-
-@tab_completions_app.command
-def uninstall(*, shell: Shell | None = None) -> None:
-    """Remove managed tab completions for the current (or specified) shell.
-
-    Deletes the hook file and cached completion script for the currently
-    active ``siesta`` executable, and strips the source line from the shell
-    RC file.  Files belonging to other siesta installations are left
-    untouched.
-
-    Parameters
-    ----------
-    shell : {'bash', 'zsh'} | None, optional
-        Target shell. Defaults to the current shell detected from ``$SHELL``.
-        Required when the current shell cannot be detected.
-    """
-    from siesta.completions import uninstall_managed_completion
-
-    resolved = _resolve_shell(shell)
-    result = uninstall_managed_completion(shell=resolved)
-    removed: list[Path] = result["removed"]
-    missing: list[Path] = result["missing"]
-
-    if removed:
-        for p in removed:
-            rprint(f"[green]Removed: {p}[/green]")
-    if missing:
-        for p in missing:
-            rprint(f"[dim]Not found (skipped): {p}[/dim]")
-    if not removed and not missing:
-        rprint("[dim]Nothing to remove.[/dim]")
-
-
-def _set_completion_hint() -> None:
-    """Show a tab-completion install tip in ``--help`` when not already installed."""
-    current_shell = detect_current_shell()
-    if current_shell is None or not is_completion_installed(current_shell):
-        shell_flag = " --shell <bash|zsh>" if current_shell is None else ""
-        app.help_epilogue = (
-            f"Tip: enable tab completions with "
-            f"`siesta self tab-completions install{shell_flag}`"
-        )
-
-
-def main():
-    """Run the CLI, gracefully handling ``KeyboardInterrupt``."""
-    _set_completion_hint()
-
-    # Start background update check (non-blocking)
-    update_future = start_background_update_check(metadata.version("siesta"))
-    interrupted = False
-
-    try:
-        app()
-    except KeyboardInterrupt:
-        interrupted = True
-        logger.abort("\nAborted.", exit=1)
-    finally:
-        # Show update message at the end (if available)
-        if not interrupted:
-            update_msg = get_update_message(update_future)
-            if update_msg:
-                logger.print(update_msg)
-
-
-@self_app.command(name="set-github-pat")
-def set_github_pat(pat: Optional[str] = ""):
-    """
-    Store a GitHub Personal Access Token (PAT) in your keyring.
-
-    A Github PAT is only required when using the ``--remote-assets`` flag with commands
-    like ``siesta docs init``, ``siesta docs update``, or ``siesta project quickstart``
-    to fetch the latest boilerplate files from the remote repository. By default, local
-    bundled files are used and no PAT is needed.
-
-    `About GitHub PAT <https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens#about-personal-access-tokens>`_
-
-    `Creating Github a PAT <https://github.com/settings/personal-access-tokens>`_
-
-
-    1. Go to ``Settings > Developer settings > Personal access tokens (fine-grained) >
-       Generate new token``.
-    2. Name it ``siesta``.
-    3. Set ``Entalpic`` as resource owner
-    4. Expire it in 1 year.
-    5. Only select the ``siesta`` repository
-    6. Set *Repository Permissions* to *Contents: Read* and *Metadata: Read*.
-    7. Click on *Generate token*.
-
-
-    Parameters
-    ----------
-    pat : str, optional
-        The GitHub Personal Access Token.
-    """
-    from keyring import set_password
-
-    assert isinstance(pat, str), "PAT must be a string."
-
-    logger.warning(
-        "Run [r]$ siesta self set-github-pat --help[/r]"
-        + " if you're not sure how to generate a PAT."
-    )
-    if not pat:
-        pat = getpass.getpass("Enter your GitHub PAT (hidden): ")
-    logger.confirm(
-        f"Are you sure you want to set the GitHub PAT to {pat[:5]}...{pat[-5:]}?"
-    )
-    set_password("siesta", "github_pat", pat)
-    logger.success(
-        "GitHub PAT set. You can now use ``--remote-assets`` to fetch remote files."
-    )
-
-
-@self_app.command(name="show-github-pat")
-def show_github_pat(full: bool = False):
-    """
-    Display the stored GitHub Personal Access Token (PAT).
-
-    By default, shows a masked version of the token (first 14 and last 4 characters)
-    for identification without exposing the full secret.
-
-    Parameters
-    ----------
-    full : bool, optional
-        Show the full token in plaintext instead of the masked version.
-    """
-    pat = get_user_pat()
-    if not pat:
-        logger.warning(
-            "No GitHub PAT found."
-            + " Run [r]$ siesta self set-github-pat[/r] to set one."
-        )
-        return
-
-    if full:
-        logger.warning("The full token will be displayed in plaintext.")
-        logger.info(f"GitHub PAT: {pat}")
-    else:
-        masked = f"{pat[:14]}...{pat[-4:]}"
-        logger.info(f"GitHub PAT: {masked}")
-        logger.info("Use [r]--full[/r] to display the entire token.")
-
-
-@self_app.command(name="show-deps")
-def show_deps(as_pip: bool = False):
-    """Show the recommended dependencies for the documentation that would be installed with `siesta docs init`.
-
-    Parameters
-    ----------
-    as_pip : bool, optional
-        Show as pip install command.
-    """
-    deps = load_deps()
-    if as_pip:
-        logger.print(" ".join([d for k in deps for d in deps[k]]))
-    else:
-        logger.print("Dependencies:")
-        for scope in deps:
-            logger.print("  • " + scope + ": " + " ".join(deps[scope]))
-
-
-@docs_app.command(name="init")
-def init_docs(
-    path: str = "./docs",
-    as_main_deps: bool = None,
-    overwrite: bool = False,
-    deps: bool = None,
-    uv: bool = None,
-    interactive: Annotated[bool, Parameter(name=["-i", "--interactive"])] = False,
-    branch: str = "main",
-    contents: str = "src/siesta/boilerplate",
-    remote_assets: bool = False,
-    project_name: str = None,
-):
-    """Initialize a Sphinx documentation project with Entalpic's standard configuration (also called within ``siesta project quickstart``).
-
-    In particular:
-
-    - Initializes a new Sphinx project at the specified path.
-
-    - Optionally installs recommended dependencies (run ``siesta self show-deps`` to see
-      them).
-
-    - Uses the split source / build folder structure.
-
-    - Includes standard `conf.py` and `index.rst` files with good defaults.
-
-    .. warning::
-
-        If you don't install the dependencies here, you will need to install them
-        manually.
-
-    .. important::
-
-        Update the placeholders (``$FILL_HERE``) in the generated files with the
-        appropriate values before you build the documentation.
-
-    .. tip::
-
-        Build the local HTML docs by running ``$ siesta docs build`` or ``$ siesta docs watch``.
-
-    Parameters
-    ----------
-    path : str, optional
-        Where to store the docs.
-    as_main_deps : bool, optional
-        Whether docs dependencies should be included in the main or dev dependencies.
-    overwrite : bool, optional
-        Whether to overwrite existing docs (if any).
-    deps : bool, optional
-        Prevent dependencies prompt by forcing its value to ``True`` or ``False``.
-    uv : bool, optional
-        Prevent uv prompt by forcing its value to ``True`` or ``False``.
-    interactive : bool, optional
-        Enable interactive mode with prompts for all options (``-i``). By default,
-        sensible defaults are used. User-specified flags always take precedence.
-    branch : str, optional
-        The branch to fetch the static files from.
-    contents : str, optional
-        The path to the static files in the repository.
-    remote_assets : bool, optional
-        Fetch boilerplate docs assets from the remote GitHub repository instead of
-        using the local bundled files. Requires a GitHub Personal Access Token (PAT).
-        Run ``$ siesta self set-github-pat`` to configure one.
-    project_name : str, optional
-        The project's name. If not provided, it will be prompted.
-    Raises
-    ------
-    sys.exit(1)
-        If the path already exists and ``--overwrite`` is not provided.
-    """
-    # Check for Python files before proceeding
-    if not has_python_files():
-        logger.abort(
-            "No Python files found in project. Documentation requires Python files to document.",
-            exit=1,
-        )
-
-    # Check for GitHub Personal Access Token (only needed for remote assets)
-    if remote_assets:
-        pat = get_user_pat()
-        if not pat:
-            logger.warning(
-                "You need to set a GitHub Personal Access Token"
-                + " to fetch the latest static files."
-            )
-            logger.warning(
-                "Run [r]$ siesta self set-github-pat --help[/r] to learn how to."
-            )
-            logger.abort("Aborting.", exit=1)
-
-    # Setting defaults: only fill in values that weren't explicitly provided
-    if not interactive:
-        if deps is None:
-            deps = CLI_DEFAULTS["deps"]
-        if as_main_deps is None:
-            as_main_deps = CLI_DEFAULTS["as_main_deps"]
-
-    # Where the docs will be stored, typically `$CWD/docs`
-    path = resolve_path(path)
-    logger.info(f"Initializing docs at path: [r]{path}[/r]")
-    if path.exists():
-        # docs folder already exists
-        if not overwrite:
-            # user doesn't want to overwrite -> abort
-            logger.warning(f"Path already exists: {path}")
-            logger.warning("Use --overwrite to overwrite.")
-            logger.abort("Aborting.", exit=1)
-
-    # Prompt collection phase: gather decisions before any mutation.
-    if deps is None:
-        deps = logger.confirm("Would you like to install recommended dependencies?")
-
-    with_uv = False
-    if deps:
-        # Check if uv.lock exists in order to decide whether to use uv or not
-        if resolve_path("./uv.lock").exists():
-            if uv is not None:
-                with_uv = uv
-            elif not interactive:  # if not interactive, assume uv since uv.lock exists
-                with_uv = True
-            else:
-                with_uv = logger.confirm(
-                    "It looks like you are using uv. Use `uv add` to add dependencies?"
-                )
-        else:
-            if uv:
-                logger.warning(
-                    "uv.lock not found. Skipping uv dependencies, installing with pip."
-                )
-
-    # Execution phase: perform mutations only after all decisions are collected.
-    if path.exists():
-        logger.warning("🚧 Overwriting path.")
-        rmtree(path)
-
-    path.mkdir(parents=True)
-    logger.success("Docs initialized 📄")
-
-    # Install dependencies if requested.
-    if deps:
-        logger.info(f"Installing docs dependencies{' with uv.' if with_uv else '.'}..")
-        # Execute the command to install dependencies
-        install_dependencies(with_uv, with_uv and not as_main_deps)
-        logger.info("Docs dependencies installed.")
-    else:
-        logger.warning("Skipping dependency installation.")
-
-    # Download and copy siesta pre-filled folder structure to the target directory
-    copy_boilerplate(
-        path,
-        branch=branch,
-        content_path=contents,
-        overwrite=True,
-        local=not remote_assets,
-    )
-    # Make empty dirs (_build and _static) in target directory
-    make_empty_folders(path)
-    # Update defaults from user config
-    overwrite_docs_files(path, interactive, project_name)
-
-    # Check if the ReadTheDocs YAML config exists
-    has_rtd = Path(".readthedocs.yaml").exists()
-    if not has_rtd:
-        write_rtd_config()
-        logger.info("ReadTheDocs config written.")
-    else:
-        logger.info("ReadTheDocs config already exists.")
-
-    logger.info(
-        "Now go to your newly created docs folder and update placeholders in"
-        + " [r]conf.py[/r] with the appropriate values.",
-    )
-
-    build_docs(path)
-
-
-@docs_app.command(name="update")
-def update(
-    path: str = "./docs",
-    branch: str = "main",
-    contents: str = "src/siesta/boilerplate",
-    remote_assets: bool = False,
-):
-    """
-    Update the static files in the docs folder like CSS, JS and images.
-
-    By default, uses the local bundled boilerplate files. Use ``--remote-assets`` to
-    fetch the latest version from the remote GitHub repository (requires a GitHub PAT;
-    run ``$ siesta self set-github-pat`` to configure one).
-
-    .. note::
-
-        Existing files will be backed up to ``{filename}.bak`` before new files are
-        copied.
-
-    Parameters
-    ----------
-    path : str, optional
-        The path to your documentation folder.
-    branch : str, optional
-        The branch to fetch the static files from.
-    contents : str, optional
-        The path to the static files in the repository.
-    remote_assets : bool, optional
-        Fetch boilerplate docs assets from the remote GitHub repository instead of
-        using the local bundled files. Requires a GitHub Personal Access Token (PAT).
-        Run ``$ siesta self set-github-pat`` to configure one.
-    """
-    # Check if the path exists
-    path = resolve_path(path)
-    if not path.exists():
-        logger.abort(f"Path not found: {path}", exit=1)
-
-    # Prompt collection phase: gather all confirmations up front.
-    update_static = logger.confirm(
-        "Overwrite the documentation's HTML static files. Continue?"
-    )
-    update_conf = logger.confirm("Would you like to update the conf.py file?")
-    update_precommit = logger.confirm("Would you like to update the pre-commit hooks?")
-
-    if not any([update_static, update_conf, update_precommit]):
-        logger.info("No updates selected. Nothing to do.")
-        logger.success("Done.")
-        return
-
-    # Execution phase: run only the selected mutating actions.
-    if update_static:
-        static = path / "source" / "_static"
-        if not static.exists():
-            logger.abort(f"Static folder not found: {static}", exit=1)
-        copy_boilerplate(
-            dest=path,
-            branch=branch,
-            content_path=contents,
-            overwrite=False,
-            include_files_regex="_static",
-            local=not remote_assets,
-        )
-        logger.success("Static files updated.")
-
-    if update_conf:
-        update_conf_py(path, branch=branch, local=not remote_assets)
-        logger.success("[r]conf.py[/r] updated.")
-
-    if update_precommit:
-        write_or_update_pre_commit_file()
-        has_uv = Path("uv.lock").exists()
-        if has_uv:
-            run_command(["uv", "add", "--dev", "pre-commit"])
-            run_command(["uv", "run", "pre-commit", "install"])
-        else:
-            run_command(["pre-commit", "install"])
-        logger.success("Pre-commit hooks updated.")
-    logger.success("Done.")
-
-
-@docs_app.command(name="build")
-def build_docs(path: str = "./docs"):
-    """Build your docs.
-
-    Equivalent to running ``$ cd docs && make clean && make html``.
-
-    Parameters
-    ----------
-    path : str, optional
-        The path to your documentation folder.
-    """
-    path = resolve_path(path)
-    if not path.exists():
-        logger.abort(f"Path not found: {path}", exit=1)
-    make = path / "Makefile"
-    if not make.exists():
-        logger.abort(f"Makefile not found in {path}.", exit=1)
-    commands = [
-        ["make", "clean"],
-        ["make", "html"],
-    ]
-    with_uv = Path("uv.lock").exists()
-    if with_uv:
-        commands = [["uv", "run"] + command for command in commands]
-    for c, command in enumerate(commands):
-        with logger.loading(f"Running {' '.join(command)}..."):
-            result = run_command(command, cwd=str(path), check=False)
-        if result.returncode != 0:
-            logger.error(result.stderr, title="Build Error", as_panel=True)
-            logger.abort("Failed to build the docs.")
-
-        if c == 0:
-            continue
-
-        logger.info(result.stdout, title="Build Output", as_panel=True)
-        if result.stderr:
-            logger.warning(
-                result.stderr,
-                title=f"Warnings executing [r]{' '.join(command)}[/r]",
-                as_panel=True,
-            )
-    logger.info(
-        "Ask Victor if you want to automatically build and deploy the docs to ReadTheDocs."
-    )
-    docs_path = (path / "build/html/index.html").relative_to(Path.cwd())
-    logger.info(f"Local docs built in {docs_path}")
-    logger.success("Open locally with [r]siesta docs open[/r]")
-
-
-@docs_app.command(name="watch")
-def watch_docs(
-    path: str = "./docs", patterns: str = r".+/src/.+\.py;.+/source/.+\.rst"
-):
-    """Automatically build the docs when source files matching the given patterns are changed.
-
-    Files must be accessible down the directory you run the command from.
-
-    Files in the ``/autoapi/`` folder are not watched so that the intermediate
-    files generated by AutoAPI do not trigger a continuous rebuild.
-
-    Parameters
-    ----------
-    path : str, optional
-        The path to your documentation folder.
-    patterns : str, optional
-        The patterns to watch for changes, separated by ``;``.
-    """
-    # Build the docs
-    build_docs(path)
-
-    # Patterns to watch
-    patterns = [p.strip() for p in patterns.split(";")]
-
-    # Watch for changes
-    abd = AutoBuildDocs(patterns, build_docs, path=path)
-    observer = Observer()
-    observer.schedule(abd, path=".", recursive=True)
-    observer.start()
-    here = Path().resolve()
-    logger.info(f"Watching {here}. Press Ctrl+C to stop.")
-    try:
-        while True:
-            time.sleep(1)
-    except KeyboardInterrupt:
-        observer.stop()
-        observer.join()
-    logger.print()
-    logger.warning("Watching stopped. Bye bye 👋")
-
-
-@docs_app.command(name="open")
-def open_docs(path: str = "./docs"):
-    """Open the locally-built docs in the default browser.
-
-    Parameters
-    ----------
-    path : str, optional
-        The path to your documentation folder.
-    """
-    path = resolve_path(path)
-    index = path / "build/html/index.html"
-    if not index.exists():
-        logger.abort(f"Index file not found: {index}", exit=1)
-
-    logger.info(f"Opening {index} in the default browser.")
-    if platform.system() == "Darwin":  # macOS
-        subprocess.call(("open", str(index)))
-    elif platform.system() == "Windows":  # Windows
-        os.startfile(str(index))
-    else:  # Linux variants
-        subprocess.call(("xdg-open", str(index)))
-
-
-@project_app.command(name="quickstart")
-def quickstart_project(
-    as_app: bool = False,
-    as_pkg: bool = False,
-    precommit: bool | None = None,
-    docs: bool | None = None,
-    deps: bool | None = None,
-    docs_path: str = "./docs",
-    as_main_deps: bool | None = None,
-    overwrite: bool = False,
-    interactive: Annotated[bool, Parameter(name=["-i", "--interactive"])] = False,
-    branch: str = "main",
-    contents: str = "src/siesta/boilerplate",
-    remote_assets: bool = False,
-    ipdb: bool | None = None,
-    tests: bool | None = None,
-    actions: bool | None = None,
-    gitignore: bool | None = None,
-):
-    """Start a ``uv``-based Python project from scratch, with initial project structure and docs.
-
-    Overall:
-
-    * Initializes a new ``uv`` project with ``$ uv init``.
-    * Installs recommended dependencies with ``$ uv add --dev [...]``.
-    * Sets up pytest testing infrastructure (via ``siesta project setup-tests``).
-    * Sets up GitHub Actions for CI.
-    * Initializes a new Sphinx project at the specified path as per ``$ siesta
-      docs init``.
-    * Initializes pre-commit hooks with ``$ uv run pre-commit install``.
-
-    .. tip::
-
-        If you only need to add testing infrastructure to an existing project,
-        use ``siesta project setup-tests`` instead.
-
-    .. note::
-
-        The default behavior is to initialize the project as a library (with a package
-        structure within a ``src/`` folder). Use the ``--as-app`` flag to initialize the
-        project as an app (just a script file to start with) or ``--as-pkg`` to initialize
-        as a package (with a package structure in the root directory).
-
-    .. important::
-
-        By default, sensible defaults are used for unspecified options:
-
-        - Install recommended dependencies.
-        - Initialize pre-commit hooks.
-        - Initialize pytest testing infrastructure and GitHub Actions.
-        - Initialize the docs.
-
-        User-specified flags always take precedence. For example,
-        ``--no-tests`` will skip test initialization.
-
-        Use ``-i`` or ``--interactive`` to be prompted for each option instead.
-
-    .. important::
-
-        If you generate the docs (with ``--docs`` or by default) parameters
-        like ``--deps`` and ``--as_main_deps`` will passed to the ``siesta docs init``
-        command so it may be worth checking ``$ siesta docs init --help``.
-
-    Parameters
-    ----------
-    as_app : bool, optional
-        Whether to initialize the project as an app (just a script file to start with).
-    as_pkg : bool, optional
-        Whether to initialize the project as a package (with a package structure in the root directory).
-    precommit : bool, optional
-        Whether to install pre-commit hooks, by default ``None`` (i.e. prompt the user).
-    docs : bool, optional
-        Whether to initialize the docs, by default ``None`` (i.e. prompt the user).
-    deps : bool, optional
-        Whether to install dependencies (dev &? docs), by default ``None`` (i.e. prompt the user).
-    docs_path : str, optional
-        Where to build the docs, by default ``./docs``.
-    as_main_deps : bool, optional
-        Whether to include docs dependencies in the main dependencies, by default
-        ``None`` (i.e. prompt the user).
-    overwrite : bool, optional
-        Whether to overwrite existing files (if any). Will be passed to ``siesta
-        docs init``.
-    interactive : bool, optional
-        Enable interactive mode with prompts for all options (``-i``). By default,
-        sensible defaults are used. User-specified flags always take precedence.
-    branch : str, optional
-        The branch to fetch the static files from.
-    contents : str, optional
-        The path to the static files in the repository.
-    remote_assets : bool, optional
-        Fetch boilerplate docs assets from the remote GitHub repository instead of
-        using the local bundled files. Requires a GitHub Personal Access Token (PAT).
-        Run ``$ siesta self set-github-pat`` to configure one.
-    ipdb: bool, optional
-        Whether to add ipdb as debugger, by default ``None`` (i.e. prompt the user).
-    tests: bool, optional
-        Whether to initialize (pytest) tests infra, by default ``None`` (i.e. prompt the user).
-    actions: bool, optional
-        Whether to initialize GitHub Actions, by default ``None`` (i.e. prompt the user).
-    gitignore: bool, optional
-        Whether to initialize the ``.gitignore`` file, by default ``None`` (i.e. prompt the user).
-    """
-    if as_app and as_pkg:
-        logger.abort("Cannot use both --as-app and --as-pkg flags.")
-
-    # uv is required
-    has_uv = bool(run_command(["uv", "--version"]))
-    if not has_uv:
-        logger.abort(
-            "uv not found. Please install it first -> https://docs.astral.sh/uv/getting-started/installation/"
-        )
-
-    # Get the project name
-    project_name = get_project_name(interactive)
-
-    # Setting defaults: only fill in values that weren't explicitly provided
-    if not interactive:
-        if precommit is None:
-            precommit = CLI_DEFAULTS["precommit"]
-        if docs is None:
-            docs = CLI_DEFAULTS["docs"]
-        if deps is None:
-            deps = CLI_DEFAULTS["deps"]
-        if as_main_deps is None:
-            as_main_deps = CLI_DEFAULTS["as_main_deps"]
-        if ipdb is None:
-            ipdb = CLI_DEFAULTS["ipdb"]
-        if tests is None:
-            tests = CLI_DEFAULTS["tests"]
-        if actions is None:
-            actions = CLI_DEFAULTS["actions"]
-        if gitignore is None:
-            gitignore = CLI_DEFAULTS["gitignore"]
-
-    # Prompt collection phase: gather all unresolved decisions before mutations.
-    if deps is None:
-        deps = logger.confirm("Would you like to install recommended dependencies?")
-
-    if precommit is None:
-        precommit = logger.confirm(
-            "Would you like to update & install pre-commit hooks?"
-        )
-
-    if ipdb is None:
-        ipdb = logger.confirm("Would you like to add ipdb as debugger?")
-
-    if tests is None:
-        tests = logger.confirm("Would you like to initialize (pytest) tests infra?")
-
-    if actions is None:
-        actions = logger.confirm("Would you like to initialize GitHub Actions?")
-
-    if gitignore is None:
-        gitignore = logger.confirm(
-            "Would you like to initialize the ``.gitignore`` file?"
-        )
-
-    if docs is None:
-        docs = logger.confirm("Would you like to initialize the docs?")
-
-    docs_with_uv: bool | None = None
-    if docs and deps:
-        if Path("uv.lock").exists():
-            if interactive:
-                docs_with_uv = logger.confirm(
-                    "It looks like you are using uv. Use `uv add` to add dependencies?"
-                )
-            else:
-                docs_with_uv = True
-
-    # Execution phase: perform mutations after all decisions are collected.
-    has_uv_lock = Path("uv.lock").exists()
-    if not has_uv_lock:
-        cmd = ["uv", "init"]
-        cmd.append(f"--name={project_name}")
-        if as_app:
-            # Initialize as app (no src/ directory)
-            pass
-        elif as_pkg:
-            # Initialize as package (package structure in root directory)
-            cmd.append("--package")
-        else:
-            # Initialize as library (package structure in src/ directory)
-            cmd.append("--lib")
-
-        initialized = run_command(cmd)
-        if initialized is False:
-            logger.abort("Failed to initialize the project.")
-        logger.info("Project initialized with uv.")
-    else:
-        logger.info("Project already initialized with uv.")
-
-    if deps:
-        dev_deps = load_deps()["dev"]
-        installed = run_command(["uv", "add", "--dev"] + dev_deps)
-        if installed is False:
-            logger.abort("Failed to install the dev dependencies.")
-        logger.info("Dev dependencies installed.")
-
-    if precommit:
-        write_or_update_pre_commit_file()
-        pre_commit_installed = run_command(["uv", "run", "pre-commit", "install"])
-        if pre_commit_installed is False:
-            logger.abort("Failed to install pre-commit hooks.")
-        logger.info("Pre-commit hooks installed.")
-
-    if ipdb:
-        add_ipdb_as_debugger()
-        logger.info("[r]ipdb[/r] added as debugger.")
-
-    if tests:
-        # Decision ownership stays in quickstart; setup-tests receives explicit values.
-        setup_tests(
-            project_name=project_name,
-            actions=actions,
-            deps=False,  # deps already handled by quickstart
-            interactive=False,
-        )
-    elif actions:
-        # Warn if user wants CI but has no tests
-        if not Path("tests").exists():
-            logger.warning(
-                "You're setting up GitHub Actions CI without tests. "
-                "Either add tests later or update [r].github/workflows/test.yml[/r] "
-                "to match your project's needs."
-            )
-        write_test_actions_config()
-        logger.info("Test actions config written.")
-
-    if gitignore:
-        write_gitignore()
-        logger.info("Gitignore written.")
-
-    if docs:
-        init_docs(
-            path=docs_path,
-            as_main_deps=as_main_deps,
-            overwrite=overwrite,
-            deps=deps,
-            uv=docs_with_uv,
-            interactive=False,
-            branch=branch,
-            contents=contents,
-            remote_assets=remote_assets,
-            project_name=project_name,
-        )
-
-    tree_project(".")
-    logger.info("Project initialized.")
-    logger.success("🔥 Happy coding! 👋")
-
-
-@project_app.command(name="setup-tests")
-def setup_tests(
-    project_name: str | None = None,
-    actions: bool | None = None,
-    deps: bool | None = None,
-    interactive: Annotated[bool, Parameter(name=["-i", "--interactive"])] = False,
-):
-    """Set up pytest testing infrastructure for an existing project.
-
-    This command adds testing infrastructure to an existing Python project:
-
-    - Installs ``pytest`` and ``pytest-cov`` as dev dependencies.
-    - Creates a ``tests/`` directory with an example test file.
-    - Optionally sets up GitHub Actions for CI (runs tests on PRs and pushes to main).
-
-    .. tip::
-
-        This is useful when you have an existing project that doesn't have tests yet,
-        or when you want to add Entalpic's standard testing setup to a project.
-
-    .. note::
-
-        This command assumes a ``uv``-based project (``uv.lock`` exists).
-        If ``uv.lock`` is not found, dependencies will be installed with ``pip``.
-
-    Example
-    -------
-    .. code-block:: bash
-
-        # Set up tests with all defaults (installs deps + GitHub Actions)
-        $ siesta project setup-tests
-
-        # Set up tests interactively (prompts for each option)
-        $ siesta project setup-tests -i
-
-        # Set up tests without GitHub Actions (user flag takes precedence)
-        $ siesta project setup-tests --no-actions
-
-        # Set up tests for a specific project name
-        $ siesta project setup-tests --project-name=myproject
-
-    Parameters
-    ----------
-    project_name : str, optional
-        The project's name. If not provided, it will be detected from ``pyproject.toml``
-        or prompted.
-    actions : bool, optional
-        Whether to initialize GitHub Actions, by default ``None`` (i.e. prompt the user
-        in interactive mode, defaults to ``True`` otherwise).
-    deps : bool, optional
-        Whether to install test dependencies (``pytest``, ``pytest-cov``), by default
-        ``None`` (i.e. prompt the user in interactive mode, defaults to ``True`` otherwise).
-    interactive : bool, optional
-        Enable interactive mode with prompts for all options (``-i``). By default,
-        sensible defaults are used. User-specified flags always take precedence.
-    """
-    # Get the project name
-    if project_name is None:
-        project_name = get_project_name(interactive)
-
-    # Setting defaults: only fill in values that weren't explicitly provided
-    if not interactive:
-        if actions is None:
-            actions = CLI_DEFAULTS["actions"]
-        if deps is None:
-            deps = CLI_DEFAULTS["deps"]
-
-    # Check if uv is available and uv.lock exists
-    has_uv = Path("uv.lock").exists()
-
-    # Prompt collection phase: gather decisions before mutations.
-    if deps is None:
-        deps = logger.confirm("Would you like to install test dependencies?")
-
-    if actions is None:
-        actions = logger.confirm("Would you like to initialize GitHub Actions?")
-
-    # Execution phase: apply selected mutations.
-    if deps:
-        test_deps = ["pytest", "pytest-cov"]
-        if has_uv:
-            installed = run_command(["uv", "add", "--dev"] + test_deps)
-            if installed is False:
-                logger.abort("Failed to install test dependencies.")
-            logger.info("Test dependencies installed with uv.")
-        else:
-            installed = run_command(["pip", "install"] + test_deps)
-            if installed is False:
-                logger.abort("Failed to install test dependencies.")
-            logger.info("Test dependencies installed with pip.")
-
-    # Write the tests infrastructure
-    write_tests_infra(project_name)
-    logger.info("Tests infra written.")
-
-    if actions:
-        write_test_actions_config()
-        logger.info("Test actions config written.")
-
-    logger.success("Testing infrastructure set up successfully.")
-
-
-@project_app.command(name="tree")
-def tree_project(path: str = ".", ignore_from_gitignore: bool = True):
-    """Show a project's tree, i.e. its directory structure.
-
-    Additionally displays labels to explain the purpose of useful files / directories.
-
-    Parameters
-    ----------
-    path : str, optional
-        The path to the project.
-    ignore_from_gitignore : bool, optional
-        Whether to ignore files from the ``.gitignore`` file in ``path``.
-        Only useful if ``path`` is a git repository.
-    """
-    path = resolve_path(path)
-    if not path.exists():
-        logger.abort(f"Path not found: {path}", exit=1)
-    gitignore_path = path / ".gitignore"
-    if ignore_from_gitignore and gitignore_path.exists():
-        ignore = parse_gitignore(gitignore_path)
-    else:
-        ignore = None
-    term_cols = get_terminal_size().columns
-    logger.info(
-        ".\n"
-        + make_labeled_tree(
-            path,
-            ignore,
-            max_line_length=term_cols - 4,  # account for panel borders
-        ),
-        as_panel=True,
-        title="Your project's directory structure",
-    )
-
-
-# ============================================================================
-# Self commands
-# ============================================================================
-
-
-@self_app.command(name="version")
-def self_version():
-    """Show the current siesta version and check for updates.
-
-    Displays the installed pip version, latest GitHub release,
-    and latest commit on main.
-    """
-    from siesta import __version__
-
-    logger.info(f"Installed (pip): [r]{__version__}[/r]")
-
-    with logger.loading("Fetching latest versions from GitHub..."):
-        release_version, release_err = get_latest_github_release_version()
-        commit_info, commit_err = get_latest_commit_info()
-
-    if release_version is not None:
-        logger.info(f"GitHub release:  [r]{release_version}[/r]")
-    else:
-        _detail = f" ({release_err})" if release_err else ""
-        logger.warning(f"GitHub release:  could not fetch{_detail}")
-
-    if commit_info is not None:
-        time_str = commit_info["time"].strftime("%Y-%m-%d %H:%M:%S UTC")
-        logger.info(
-            f"GitHub main:     [r]{commit_info['hash']}[/r]"
-            f" by {commit_info['author']}"
-            f" ({time_str})"
-        )
-    else:
-        _detail = f" ({commit_err})" if commit_err else ""
-        logger.warning(f"GitHub main:     could not fetch{_detail}")
-
-    if release_version is not None:
-        comparison = compare_versions(__version__, release_version)
-        if comparison < 0:
-            logger.warning(f"A newer version is available: [r]{release_version}[/r]")
-            logger.info("Run [r]siesta self update[/r] to upgrade.")
-        elif comparison > 0:
-            logger.info("You are running a pre-release or development version.")
-        else:
-            logger.success("You are running the latest release.")
-
-
-@self_app.command(name="update")
-def self_update(force: bool = False, dry: bool = False):
-    """Update siesta to the latest version.
-
-    Automatically detects how siesta was installed (uv tool, pipx, pip, or editable)
-    and uses the appropriate update command (uv tool, pipx, pip), except for editable
-    installations (i.e. when siesta is installed from source) in which case it
-    shows a warning and suggests manual steps.
-
-    Parameters
-    ----------
-    force : bool, optional
-        Force update even if already on the latest version.
-    dry : bool, optional
-        Show what would be done without actually updating.
-    """
-    from siesta import __version__
-
-    # Get installation method
-    method = get_installation_method()
-    method_display = {"uv": "uv tool", "pipx": "pipx", "pip": "pip"}.get(method, method)
-
-    # Handle editable installations
-    if method == "editable":
-        logger.warning("siesta is installed in editable (development) mode.")
-        logger.info("To update, navigate to the source directory and run:")
-        logger.info("  [r]git pull[/r]")
-        logger.info("  [r]uv pip install -e .[/r]  (or pip install -e .)")
-        return
-
-    # Dry run: just show what would be done
-    if dry:
-        cmd = get_update_command(method)
-        logger.info(f"Installation method: [r]{method_display}[/r]")
-        logger.info(f"Would run: [r]{' '.join(cmd)}[/r]")
-        return
-
-    # Check current vs latest version
-    with logger.loading("Checking for updates..."):
-        latest, check_err = get_latest_version()
-
-    if latest is None:
-        _detail = f" ({check_err})" if check_err else ""
-        logger.warning(f"Could not check for latest version{_detail}.")
-        if not force:
-            logger.info("Use [r]--force[/r] to update anyway.")
-            return
-
-    if latest and not force:
-        comparison = compare_versions(__version__, latest)
-        if comparison >= 0:
-            logger.success(f"Already up to date (version {__version__}).")
-            return
-
-    # Perform the update
-    logger.info(f"Updating siesta via {method_display}...")
-
-    success = update_siesta(method)
-    if success:
-        logger.success("siesta has been updated successfully.")
-        logger.info(
-            "Restart your terminal or run [r]siesta self version[/r] to verify."
-        )
-    else:
-        logger.error("Update failed.")
-        logger.info("Try updating manually:")
-        if method == "uv":
-            logger.info("  [r]uv tool upgrade siesta[/r]")
-        elif method == "pipx":
-            logger.info("  [r]pipx upgrade siesta[/r]")
-        else:
-            logger.info("  [r]pip install --upgrade siesta[/r]")
-
-
-# Register 'upgrade' as an alias for 'update'
-self_app.command(self_update, name="upgrade")
