+++ src/siesta/cli/self_app.py
@@ -1,0 +1,449 @@
+# Copyright 2025 Entalpic
+"""Self-management and tab-completion CLI commands."""
+
+import builtins
+import getpass
+from pathlib import Path
+from textwrap import dedent
+from typing import Optional
+
+from cyclopts import App
+from rich import print as rprint
+
+from siesta.cli._shared import resolve_shell
+from siesta.completions import Shell
+from siesta.utils.common import load_deps, logger
+from siesta.utils.github import (
+    get_latest_commit_info,
+    get_latest_github_release_version,
+    get_user_pat,
+)
+from siesta.utils.self import (
+    compare_versions,
+    get_installation_method,
+    get_latest_version,
+    get_update_command,
+    update_siesta,
+)
+
+self_app = App(
+    name="self",
+    help=dedent(
+        """
+        Manage siesta itself and configure settings.
+
+        Includes version checking, updates, GitHub PAT configuration, and dependency info.
+
+        """.strip(),
+    ),
+)
+""":py:class:`cyclopts.App`: The app for the ``siesta self`` sub-command."""
+
+tab_completions_app = App(
+    name="tab-completions",
+    help="Command group — manage shell tab completions (requires a subcommand).",
+)
+self_app.command(tab_completions_app)
+
+
+@tab_completions_app.command
+def install(
+    *,
+    shell: Shell | None = None,
+    add_to_startup: bool = True,
+) -> None:
+    """Install tab completions for the current (or specified) shell.
+
+    Writes managed hook and static completion script files to the
+    shell-specific XDG config path, and optionally adds a source line to the
+    shell RC file.
+
+    Parameters
+    ----------
+    shell : {'bash', 'zsh'} | None, optional
+        Target shell. Defaults to the current shell detected from ``$SHELL``.
+        Required when the current shell cannot be detected.
+    add_to_startup : bool, optional
+        When ``True`` (default), append a source line to ``~/.bashrc`` or
+        ``~/.zshrc`` so completions activate in every new shell session.
+        Use ``--no-add-to-startup`` to write the files without editing the RC
+        file.
+    """
+    from siesta.completions import install_managed_completion
+
+    resolved = resolve_shell(shell)
+    path = install_managed_completion(
+        shell=resolved,
+        generate_fn=_root_app().generate_completion,
+        add_to_startup=add_to_startup,
+    )
+    rprint(f"[green]Managed completion installed to {path}[/green]")
+    if add_to_startup:
+        rc = ".bashrc" if resolved == "bash" else ".zshrc"
+        rprint(
+            f"[dim]Restart your shell or run `source ~/{rc}` "
+            "to activate completions now.[/dim]"
+        )
+
+
+@tab_completions_app.command
+def show(*, shell: Shell | None = None) -> None:
+    """Print the tab completion script to stdout (without installing).
+
+    Outputs the raw completion script for the given shell; suitable for
+    manual sourcing or inspection.
+
+    Parameters
+    ----------
+    shell : {'bash', 'zsh'} | None, optional
+        Target shell. Defaults to the current shell detected from ``$SHELL``.
+        Required when the current shell cannot be detected.
+    """
+    resolved = resolve_shell(shell)
+    builtins.print(_root_app().generate_completion(shell=resolved), end="")
+
+
+@tab_completions_app.command
+def where(*, shell: Shell | None = None, simple: bool = False) -> None:
+    """Display where managed completion files are stored.
+
+    By default shows a table with the path, existence status, and purpose of
+    each managed file. Use ``--simple`` for plain-text output (one path per
+    line), which is handy for scripting.
+
+    Parameters
+    ----------
+    shell : {'bash', 'zsh'} | None, optional
+        Target shell. Defaults to the current shell detected from ``$SHELL``.
+        Required when the current shell cannot be detected.
+    simple : bool, optional
+        When ``True``, print one ``<label>: <path>`` line per file with no
+        table formatting or descriptions.
+    """
+    from siesta.completions import (
+        managed_completion_paths,
+        resolve_cli_executable,
+        stable_exec_id,
+    )
+
+    resolved = resolve_shell(shell)
+    exec_path = resolve_cli_executable()
+    exec_id = stable_exec_id(exec_path)
+    paths = managed_completion_paths(shell=resolved, exec_id=exec_id)
+
+    _ENTRIES = [
+        (
+            "base_dir",
+            "Base directory",
+            "Root folder for all managed completion files for this shell.",
+        ),
+        (
+            "hook_file",
+            "Hook file",
+            f"Sourced by ~/{'.bashrc' if resolved == 'bash' else '.zshrc'} "
+            "at shell startup. Sources the static script and wraps the CLI command.",
+        ),
+        (
+            "static_file",
+            "Static completion script",
+            "Cached completion script generated by the CLI framework. "
+            "Sourced instantly at startup — no Python spawned. Its mtime "
+            "tracks the installed CLI version for staleness detection.",
+        ),
+    ]
+
+    if simple:
+        for key, label, _ in _ENTRIES:
+            p: Path = paths[key]
+            exists = "[green]exists[/green]" if p.exists() else "[dim]missing[/dim]"
+            rprint(f"[bold]{label}:[/bold] {p}  ({exists})")
+        return
+
+    from rich.table import Table
+
+    table = Table(
+        title=f"Managed completion paths — [cyan]{resolved}[/cyan]",
+        show_lines=True,
+    )
+    table.add_column("File", style="bold", no_wrap=True)
+    table.add_column("Status", no_wrap=True)
+    table.add_column("Path", no_wrap=True)
+    table.add_column("Description")
+
+    for key, label, description in _ENTRIES:
+        p = paths[key]
+        status = "[green]exists[/green]" if p.exists() else "[dim]missing[/dim]"
+        table.add_row(label, status, str(p), description)
+
+    rprint(table)
+
+
+@tab_completions_app.command
+def uninstall(*, shell: Shell | None = None) -> None:
+    """Remove managed tab completions for the current (or specified) shell.
+
+    Deletes the hook file and cached completion script for the currently
+    active ``siesta`` executable, and strips the source line from the shell
+    RC file.  Files belonging to other siesta installations are left
+    untouched.
+
+    Parameters
+    ----------
+    shell : {'bash', 'zsh'} | None, optional
+        Target shell. Defaults to the current shell detected from ``$SHELL``.
+        Required when the current shell cannot be detected.
+    """
+    from siesta.completions import uninstall_managed_completion
+
+    resolved = resolve_shell(shell)
+    result = uninstall_managed_completion(shell=resolved)
+    removed: list[Path] = result["removed"]
+    missing: list[Path] = result["missing"]
+
+    if removed:
+        for p in removed:
+            rprint(f"[green]Removed: {p}[/green]")
+    if missing:
+        for p in missing:
+            rprint(f"[dim]Not found (skipped): {p}[/dim]")
+    if not removed and not missing:
+        rprint("[dim]Nothing to remove.[/dim]")
+
+
+def _root_app():
+    """Return the root CLI app (lazy import avoids circular dependencies)."""
+    from siesta.cli.main_app import app
+
+    return app
+
+
+@self_app.command(name="set-github-pat")
+def set_github_pat(pat: Optional[str] = ""):
+    """
+    Store a GitHub Personal Access Token (PAT) in your keyring.
+
+    A Github PAT is only required when using the ``--remote-assets`` flag with commands
+    like ``siesta docs init``, ``siesta docs update``, or ``siesta project quickstart``
+    to fetch the latest boilerplate files from the remote repository. By default, local
+    bundled files are used and no PAT is needed.
+
+    `About GitHub PAT <https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens#about-personal-access-tokens>`_
+
+    `Creating Github a PAT <https://github.com/settings/personal-access-tokens>`_
+
+
+    1. Go to ``Settings > Developer settings > Personal access tokens (fine-grained) >
+       Generate new token``.
+    2. Name it ``siesta``.
+    3. Set ``Entalpic`` as resource owner
+    4. Expire it in 1 year.
+    5. Only select the ``siesta`` repository
+    6. Set *Repository Permissions* to *Contents: Read* and *Metadata: Read*.
+    7. Click on *Generate token*.
+
+
+    Parameters
+    ----------
+    pat : str, optional
+        The GitHub Personal Access Token.
+    """
+    from keyring import set_password
+
+    assert isinstance(pat, str), "PAT must be a string."
+
+    logger.warning(
+        "Run [r]$ siesta self set-github-pat --help[/r]"
+        + " if you're not sure how to generate a PAT."
+    )
+    if not pat:
+        pat = getpass.getpass("Enter your GitHub PAT (hidden): ")
+    logger.confirm(
+        f"Are you sure you want to set the GitHub PAT to {pat[:5]}...{pat[-5:]}?"
+    )
+    set_password("siesta", "github_pat", pat)
+    logger.success(
+        "GitHub PAT set. You can now use ``--remote-assets`` to fetch remote files."
+    )
+
+
+@self_app.command(name="show-github-pat")
+def show_github_pat(full: bool = False):
+    """
+    Display the stored GitHub Personal Access Token (PAT).
+
+    By default, shows a masked version of the token (first 14 and last 4 characters)
+    for identification without exposing the full secret.
+
+    Parameters
+    ----------
+    full : bool, optional
+        Show the full token in plaintext instead of the masked version.
+        Requires explicit confirmation because the token may appear in terminal
+        history, shell logs, and CI output.
+    """
+    pat = get_user_pat()
+    if not pat:
+        logger.warning(
+            "No GitHub PAT found."
+            + " Run [r]$ siesta self set-github-pat[/r] to set one."
+        )
+        return
+
+    if full:
+        logger.warning(
+            "The full token will be displayed in plaintext and may be captured "
+            "in terminal history, shell logs, or CI artifacts."
+        )
+        if not logger.confirm(
+            "Display the full GitHub PAT in plaintext? Type Y to continue"
+        ):
+            logger.info("Full PAT display cancelled. Showing masked token instead.")
+            masked = f"{pat[:14]}...{pat[-4:]}"
+            logger.info(f"GitHub PAT: {masked}")
+            return
+        logger.info(f"GitHub PAT: {pat}")
+    else:
+        masked = f"{pat[:14]}...{pat[-4:]}"
+        logger.info(f"GitHub PAT: {masked}")
+        logger.info(
+            "Use [r]--full[/r] to display the entire token (requires confirmation)."
+        )
+
+
+@self_app.command(name="show-deps")
+def show_deps(as_pip: bool = False):
+    """Show the recommended dependencies for the documentation that would be installed with `siesta docs init`.
+
+    Parameters
+    ----------
+    as_pip : bool, optional
+        Show as pip install command.
+    """
+    deps = load_deps()
+    if as_pip:
+        logger.print(" ".join([d for k in deps for d in deps[k]]))
+    else:
+        logger.print("Dependencies:")
+        for scope in deps:
+            logger.print("  • " + scope + ": " + " ".join(deps[scope]))
+
+
+@self_app.command(name="version")
+def self_version():
+    """Show the current siesta version and check for updates.
+
+    Displays the installed pip version, latest GitHub release,
+    and latest commit on main.
+    """
+    from siesta import __version__
+
+    logger.info(f"Installed (pip): [r]{__version__}[/r]")
+
+    with logger.loading("Fetching latest versions from GitHub..."):
+        release_version, release_err = get_latest_github_release_version()
+        commit_info, commit_err = get_latest_commit_info()
+
+    if release_version is not None:
+        logger.info(f"GitHub release:  [r]{release_version}[/r]")
+    else:
+        _detail = f" ({release_err})" if release_err else ""
+        logger.warning(f"GitHub release:  could not fetch{_detail}")
+
+    if commit_info is not None:
+        time_str = commit_info["time"].strftime("%Y-%m-%d %H:%M:%S UTC")
+        logger.info(
+            f"GitHub main:     [r]{commit_info['hash']}[/r]"
+            f" by {commit_info['author']}"
+            f" ({time_str})"
+        )
+    else:
+        _detail = f" ({commit_err})" if commit_err else ""
+        logger.warning(f"GitHub main:     could not fetch{_detail}")
+
+    if release_version is not None:
+        comparison = compare_versions(__version__, release_version)
+        if comparison < 0:
+            logger.warning(f"A newer version is available: [r]{release_version}[/r]")
+            logger.info("Run [r]siesta self update[/r] to upgrade.")
+        elif comparison > 0:
+            logger.info("You are running a pre-release or development version.")
+        else:
+            logger.success("You are running the latest release.")
+
+
+@self_app.command(name="update")
+def self_update(force: bool = False, dry: bool = False):
+    """Update siesta to the latest version.
+
+    Automatically detects how siesta was installed (uv tool, pipx, pip, or editable)
+    and uses the appropriate update command (uv tool, pipx, pip), except for editable
+    installations (i.e. when siesta is installed from source) in which case it
+    shows a warning and suggests manual steps.
+
+    Parameters
+    ----------
+    force : bool, optional
+        Force update even if already on the latest version.
+    dry : bool, optional
+        Show what would be done without actually updating.
+    """
+    from siesta import __version__
+
+    # Get installation method
+    method = get_installation_method()
+    method_display = {"uv": "uv tool", "pipx": "pipx", "pip": "pip"}.get(method, method)
+
+    # Handle editable installations
+    if method == "editable":
+        logger.warning("siesta is installed in editable (development) mode.")
+        logger.info("To update, navigate to the source directory and run:")
+        logger.info("  [r]git pull[/r]")
+        logger.info("  [r]uv pip install -e .[/r]  (or pip install -e .)")
+        return
+
+    # Dry run: just show what would be done
+    if dry:
+        cmd = get_update_command(method)
+        logger.info(f"Installation method: [r]{method_display}[/r]")
+        logger.info(f"Would run: [r]{' '.join(cmd)}[/r]")
+        return
+
+    # Check current vs latest version
+    with logger.loading("Checking for updates..."):
+        latest, check_err = get_latest_version()
+
+    if latest is None:
+        _detail = f" ({check_err})" if check_err else ""
+        logger.warning(f"Could not check for latest version{_detail}.")
+        if not force:
+            logger.info("Use [r]--force[/r] to update anyway.")
+            return
+
+    if latest and not force:
+        comparison = compare_versions(__version__, latest)
+        if comparison >= 0:
+            logger.success(f"Already up to date (version {__version__}).")
+            return
+
+    # Perform the update
+    logger.info(f"Updating siesta via {method_display}...")
+
+    success = update_siesta(method)
+    if success:
+        logger.success("siesta has been updated successfully.")
+        logger.info(
+            "Restart your terminal or run [r]siesta self version[/r] to verify."
+        )
+    else:
+        logger.error("Update failed.")
+        logger.info("Try updating manually:")
+        if method == "uv":
+            logger.info("  [r]uv tool upgrade siesta[/r]")
+        elif method == "pipx":
+            logger.info("  [r]pipx upgrade siesta[/r]")
+        else:
+            logger.info("  [r]pip install --upgrade siesta[/r]")
+
+
+# Register 'upgrade' as an alias for 'update'
+self_app.command(self_update, name="upgrade")
