+++ tests/test_cli/test_init_docs.py
@@ -1,8 +1,8 @@
 # Copyright 2025 Entalpic
 import pytest
 
-import siesta.cli as cli
-from siesta.cli import app
+import siesta.cli.docs_app as cli
+from siesta.cli.main_app import app
 
 
 def test_init_docs_basic(temp_project_with_git_and_remote, monkeypatch):
@@ -166,7 +166,7 @@ def test_init_docs_collects_prompts_before_mutation(
         events.append(f"confirm:{message}")
         return True
 
-    monkeypatch.setattr(cli.logger, "confirm", fake_confirm)
+    monkeypatch.setattr("siesta.cli.docs_app.logger.confirm", fake_confirm)
     monkeypatch.setattr(
         cli, "install_dependencies", lambda *_args, **_kwargs: events.append("install")
     )
@@ -177,12 +177,16 @@ def test_init_docs_collects_prompts_before_mutation(
         cli, "make_empty_folders", lambda *_args, **_kwargs: events.append("empty")
     )
     monkeypatch.setattr(
-        cli, "overwrite_docs_files", lambda *_args, **_kwargs: events.append("overwrite")
+        cli,
+        "overwrite_docs_files",
+        lambda *_args, **_kwargs: events.append("overwrite"),
     )
     monkeypatch.setattr(
         cli, "write_rtd_config", lambda *_args, **_kwargs: events.append("rtd")
     )
-    monkeypatch.setattr(cli, "build_docs", lambda *_args, **_kwargs: events.append("build"))
+    monkeypatch.setattr(
+        cli, "build_docs", lambda *_args, **_kwargs: events.append("build")
+    )
 
     orig_mkdir = cli.Path.mkdir
 
@@ -196,8 +200,12 @@ def test_init_docs_collects_prompts_before_mutation(
         app(["docs", "init", "-i"])
     assert exc_info.value.code == 0
 
-    prompt_indices = [i for i, event in enumerate(events) if event.startswith("confirm:")]
-    first_mutation_idx = next(i for i, event in enumerate(events) if not event.startswith("confirm:"))
+    prompt_indices = [
+        i for i, event in enumerate(events) if event.startswith("confirm:")
+    ]
+    first_mutation_idx = next(
+        i for i, event in enumerate(events) if not event.startswith("confirm:")
+    )
     assert prompt_indices
     assert max(prompt_indices) < first_mutation_idx
 
@@ -209,14 +217,15 @@ def test_init_docs_cancel_during_prompt_has_no_mutation(
     monkeypatch.chdir(temp_project_with_git_and_remote)
 
     monkeypatch.setattr(
-        cli.logger,
-        "confirm",
+        "siesta.cli.docs_app.logger.confirm",
         lambda _message: (_ for _ in ()).throw(KeyboardInterrupt()),
     )
     monkeypatch.setattr(
         cli.Path,
         "mkdir",
-        lambda *_args, **_kwargs: pytest.fail("mkdir should not be called on cancellation"),
+        lambda *_args, **_kwargs: pytest.fail(
+            "mkdir should not be called on cancellation"
+        ),
     )
     monkeypatch.setattr(
         cli,
