+++ tests/test_cli/test_show_deps.py
@@ -1,5 +1,5 @@
 # Copyright 2025 Entalpic
-from siesta.cli import app
+from siesta.cli.main_app import app
 from siesta.utils.common import load_deps
