+++ tests/test_cli/test_show_github_pat.py
@@ -1,7 +1,7 @@
 # Copyright 2025 Entalpic
 from unittest.mock import patch
 
-from siesta.cli import app
+from siesta.cli.main_app import app
 
 
 def test_show_github_pat_masked(capture_output):
@@ -20,23 +20,39 @@ def test_show_github_pat_masked(capture_output):
 
 
 def test_show_github_pat_full(capture_output):
-    """Test that show-github-pat --full displays the full token with a warning."""
-    with capture_output() as output:
-        try:
-            app(["self", "show-github-pat", "--full"])
-        except SystemExit as e:
-            assert e.code == 0
+    """Test that show-github-pat --full displays the full token after confirmation."""
+    with patch("siesta.cli.self_app.logger.confirm", return_value=True):
+        with capture_output() as output:
+            try:
+                app(["self", "show-github-pat", "--full"])
+            except SystemExit as e:
+                assert e.code == 0
 
     output_text = output.getvalue()
     assert "fake-github-pat-for-testing" in output_text
     assert "plaintext" in output_text
 
 
+def test_show_github_pat_full_cancelled(capture_output):
+    """Test that cancelling --full confirmation shows masked token instead."""
+    with patch("siesta.cli.self_app.logger.confirm", return_value=False):
+        with capture_output() as output:
+            try:
+                app(["self", "show-github-pat", "--full"])
+            except SystemExit as e:
+                assert e.code == 0
+
+    output_text = output.getvalue()
+    assert "fake-github-pat-for-testing" not in output_text
+    assert "fake-github-" in output_text
+    assert "cancelled" in output_text.lower()
+
+
 def test_show_github_pat_none(capture_output):
     """Test that show-github-pat warns when no PAT is stored."""
     with (
         patch("siesta.utils.github.get_user_pat", return_value=None),
-        patch("siesta.cli.get_user_pat", return_value=None),
+        patch("siesta.cli.self_app.get_user_pat", return_value=None),
     ):
         with capture_output() as output:
             try:
